Fix #1 by hardcoding command file names into a list
Dealing with resolving file names from a resource directory in vanilla java is fairly complicated, and requires separate handling for IDE and jar packaging. For such a simple demo project, I find that it is overcomplicated and would further confuse people. Instead, we will just hard code it into a visible list, and let developers determine if they want to venture into SpringBoot or finding how to do this programmatically themselves